rearrange the file structure in the js application
