+++ js/lib/config-json.js
@@ -1,0 +1,5 @@
+// put all the get json data methods here
+
+const jsonData = require('../../share/config.json')
+
+exports.jsonConfig = jsonData
