+++ js/lib/math.js
