+++ js/lib/rearrange-letters.js
